--- a/.travis.yml
+++ b/.travis.yml
@@ -4,6 +4,7 @@ python:
 before_install:
   - export DISPLAY=:99.0
   - sh -e /etc/init.d/xvfb start
+  - pip install pytest-benchmark
 install:
 - pip install .
 script: