--- a/photon_stream/tests/test_types.py
+++ b/photon_stream/tests/test_types.py
@@ -4,6 +4,7 @@ import datetime as dt
 import photon_stream as ps
 import pkg_resources
 import gzip
+import numpy as np
 
 
 def type_check(event):
@@ -17,14 +18,17 @@ def type_check(event):
     assert event.photon_stream.saturated_pixels.dtype == np.uint16
 
     for time_line in event.photon_stream.time_lines:
-        assert isinstance(time_line, array)
-        assert time_line.typecode == 'B'
+        assert isinstance(time_line, (np.ndarray, array))
+        if isinstance(time_line, np.ndarray):
+            assert time_line.dtype == np.uint8
+        else:
+            assert time_line.typecode == 'B'
 
     if hasattr(event, 'observation_info'):
         assert isinstance(event.observation_info.night, np.uint32)
         assert isinstance(event.observation_info.run, np.uint32)
         assert isinstance(event.observation_info.event, np.uint32)
-        
+
         assert isinstance(event.observation_info._time_unix_s, np.uint32)
         assert isinstance(event.observation_info._time_unix_us, np.uint32)
         assert isinstance(event.observation_info.time, dt.datetime)
@@ -46,16 +50,16 @@ def type_check(event):
 
 def test_types_from_json():
     run_path = pkg_resources.resource_filename(
-        'photon_stream', 
+        'photon_stream',
         'tests/resources/20170119_229_pass4_100events.phs.jsonl.gz')
     run = ps.EventListReader(run_path)
     event = next(run)
     type_check(event)
-    
+
 
 def test_types_from_binary():
     run_path = pkg_resources.resource_filename(
-        'photon_stream', 
+        'photon_stream',
         'tests/resources/20170119_229_pass4_100events.phs.bin.gz')
     with gzip.open(run_path, 'rb') as f:
         event = ps.experimental.io.read_event_from_file(f)