I was playing a bit with the speed of the `experimental.io.Run` speed. 

master: 733 +- 3ms
this branch: 220 +- 2 ms

Do:

    pip install pytest-benchmark
    py.test -k-slow

To get a benchmark on binary reading speed on your machine in < 20sec. 
Note the `-k-slow` parameter for pytest, the muon test takes ~80 seconds is skipped like this