@dneise 2017-08-08
Current state json is still 25% .. 30% faster than binary reader.

![2017-08-08-135815_1920x1080_scrot](https://user-images.githubusercontent.com/8200858/29070976-d41a6df2-7c41-11e7-822b-af2b4cd31c74.png)