@relleums 2018-01-25
This pull request is obsolete. The key feature of binary io speed was partly adopted already. The benchmark feature in py.test is nice and might be considered for future use